Remove static icons from the accessible name
If an icon is nested within an interactive element, it should not have an `aria-label` and the name of the icon should not be part of the accessible name of the interactive element.

Furthermore, the additional component property `label` allows to override the accessible name if it has been explicitly set. This allows to assign meaningful names to interactive elements. For example, instead of `su-times` the name of the button can now be set to `Clear input`.